put the caching stuff in a separate file
